pyproject.toml
# see https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pathnorm"
version= "0"
authors = [
    {name = "Antoine Gonon", email = "antoine.gonon@epfl.ch"},
]
description = "Compute path-norms in PyTorch"
readme = "README.md"
requires-python = ">=3.7"
keywords = ["one", "two"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "torch",
    ]

[project.optional-dependencies]
quickstart = ["torchvision"]
repro = [
    "datasets",
    "huggingface_hub",
    "ipykernel",
    "matplotlib",
    "numpy",
    "pandas",
    "pillow",
    "tensorboard",
    "torchvision",
    "tqdm",
]

[tool.setuptools]
py-modules = ["pathnorm"]

# [project.scripts]
# see https://setuptools.pypa.io/en/latest/userguide/entry_point.html#console-scripts

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
